requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.187
botocore==1.12.187
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
easydict==1.9
einops==0.1.0
entrypoints==0.3
fire==0.1.3
gast==0.2.2
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
idna==2.8
inflection==0.3.1
ipdb==0.12
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
msgpack==0.6.1
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
neovim==0.3.1
nltk==3.4.4
notebook==5.7.8
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
pynvim==0.3.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytorch-pretrained-bert==0.6.2
pytz==2019.2
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
regex==2017.4.5
requests==2.22.0
s3transfer==0.2.1
sacred==0.7.5
scipy==1.1.0
Send2Trash==1.5.0
sentencepiece==0.1.82
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.7
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.3.0
torchtext==0.3.1
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
git+https://github.com/salaniz/pycocoevalcap
git+https://github.com/huggingface/transformers@155c782a3ccd103cf63ad48a2becd7c76a7d2115#egg=transformers
entmax
